requirements.txt
discord.py>=2.4.0
feedparser>=6.0.11
requests>=2.32.3
openai>=1.65.0
mysql-connector-python>=9.2.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.34.0
